수수료 모델 교체: 매수 0 / 매도 시 매도금액 0.2%를 예수금+실현손익 양쪽 차감 + 9/8 매매일지 반영
- apply_transaction: 매수 수수료 제거, 매도 fee = 매도금액 × fee_rate 를
  예수금과 그 건 realized 양쪽에서 차감 (§6-4 "실현손익 부풀림" 오차 해소)
- account_state/checkpoint fee_rate 0.000579 → 0.002 (메리츠 국내주식 거래세)
- 체크포인트 삭제 후 전체 재생: 매도 156건 실현손익 재기록, 누적 410,291 → 405,218원
- 2026-09-08 매매일지 반영 (신규 15건): 보유 70종목, 예수금 3,675,544, 총자산 10,122,979
- 물타기 그래프 x축 span 계산 int 캐스팅(numpy timedelta DeprecationWarning 제거)
- 회귀 테스트 test_sell_tax_deducts_from_both_cash_and_realized 추가 (62 pass)

+++ ui_portfolio_tab.py
@@ -113,9 +113,9 @@ def _render_holding_detail(r: dict, tx: pd.DataFrame, T: dict):
         + (list(sells["날짜"]) if not sells.empty else []), errors="coerce")
     _xs = _xs.dropna()
     _xmin, _xmax = _xs.min(), _xs.max()
-    _span = max((_xmax - _xmin).days, 1)
-    _pad = pd.Timedelta(days=max(1, round(_span * 0.08)))
-    _dtick_ms = max(1, round(_span / 4)) * 86_400_000
+    _span = max(int((_xmax - _xmin).days), 1)
+    _pad = pd.Timedelta(days=max(1, int(round(_span * 0.08))))
+    _dtick_ms = max(1, int(round(_span / 4))) * 86_400_000
     _xrange = [(_xmin - _pad).strftime("%Y-%m-%d"), (_xmax + _pad).strftime("%Y-%m-%d")]
 
     fig = go.Figure()
